+++ core/src/task.cpp
@@ -48,33 +48,33 @@ void Task::setName(std::string name)
 	this->name = name;
 }
 
-unsigned int Task::getEarliestStartTime() const
+uint64_t Task::getEarliestStartTime() const
 {
     return earliestStartTime;
 }
-void Task::setEarliestStartTime(unsigned int earliestStartTime)
+void Task::setEarliestStartTime(uint64_t earliestStartTime)
 {
     this->earliestStartTime = earliestStartTime;
     validateAndCorrectData();
 }
-unsigned int Task::getLatestEndTime() const
+uint64_t Task::getLatestEndTime() const
 {
     return latestEndTime;
 }
-void Task::setLatestEndTime(unsigned int latestEndTime)
+void Task::setLatestEndTime(uint64_t latestEndTime)
 {
     this->latestEndTime = latestEndTime;
     validateAndCorrectData();
 }
-unsigned int Task::getLatestStartTime() const
+uint64_t Task::getLatestStartTime() const
 {
     return latestEndTime - duration;
 }
-unsigned int Task::getDuration() const
+uint64_t Task::getDuration() const
 {
     return duration;
 }
-void Task::setDuration(unsigned int duration)
+void Task::setDuration(uint64_t duration)
 {
     this->duration = duration;
     validateAndCorrectData();
@@ -107,17 +107,17 @@ std::weak_ptr<Task> Task::getParent()
     return parent;
 }
 
-unsigned int Task::getRecurringPeriod() const
+uint64_t Task::getRecurringPeriod() const
 {
 	return recurringPeriod;
 }
 
-unsigned int Task::getRecurringLateOffset() const
+uint64_t Task::getRecurringLateOffset() const
 {
 	return recurringLateOffset;
 }
 
-void Task::setRecurranceParameters(unsigned int period, unsigned int lateOffset)
+void Task::setRecurranceParameters(uint64_t period, uint64_t lateOffset)
 {
     recurringPeriod = period;
     recurringLateOffset = lateOffset;
